refactor: move OllamaRecorder to helpers module
